Execute each spec in its own monitor instance
For each spec, spawn a new instance, open the context path to that spec,
run and die or times out and die.
This is to ensure that the spec file does not helps to create brittle
specs, only production will do that.

Change the id for the root context (constant zero now), because it's
required to ignore this context sometimes (maybe in future the root context will
be removed).

Update the sample to have a spec that forgets to finish itself.