Allow configuration for contexts/examples
Using Suite#config it is possible to configure an example or every
example in a context with a configuration string.
The resulting configuration is a cascading merge from the most specific
configuration to the global (example config overwrite the surrounding
context, that overwrites the surrounding context ... that overwrites the
global configuration).