Publish and record extra info in the results
Suite#info should be used only in example blocks, but it can be
anywhere, any extra info sent will be recorded in the running example,
in practice, it means that examples get all info from ancestors
(contexts).

ConsoleReporter is printing this information, but it only makes sense if
it had a verbose report.